feat: add composable entity selector
